Types cannot subtype themselves

+++ concept/type/impl/TypeImpl.java
@@ -108,6 +108,9 @@ public abstract class TypeImpl implements Type {
     }
 
     void superTypeVertex(TypeVertex superTypeVertex) {
+        if (vertex.equals(superTypeVertex)) {
+            throw new HypergraphException(Error.TypeWrite.ILLEGAL_SUBTYPING_SELF.format(vertex.label()));
+        }
         vertex.outs().edge(Schema.Edge.Type.SUB, ((TypeImpl) sup()).vertex).delete();
         vertex.outs().put(Schema.Edge.Type.SUB, superTypeVertex);
     }
